examples/agents/requirements.txt
# Dependencies for the three demo agents (agent_sdk.py, agent_sdk1.py,
# agent_sdk2.py). Baked into the agent container image at build time.

# The aMaze SDK (sdk/ package) is installed by the Dockerfile as a separate
# step so that a single shared build context at the repo root can reach
# both the SDK and the agent code. Keep it out of this file — editable
# installs with relative paths don't survive a `docker build`.

# --- web framework for the /invoke + /chat handlers ---
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2

# --- LangChain + LangChain-OpenAI + tool plumbing ---
# langchain >=1.0 is required for `create_agent` (used by agent_sdk*.py);
# earlier 0.3.x betas only expose `create_react_agent` under
# langgraph.prebuilt.
langchain>=1.0
langchain-openai>=1.0
langchain-core>=1.0
langchain-community>=0.3
langchain-mcp-adapters>=0.1

# --- observability (optional, agent_sdk*.py imports langsmith) ---
langsmith>=0.1

# --- OpenAI SDK (langchain-openai pulls it transitively; pinned for clarity) ---
openai>=1.54
